// client/ui/src/lib/protocol/Channel.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { ChannelId } from "./ChannelId";
import type { ChannelKind } from "./ChannelKind";
import type { ServerId } from "./ServerId";

export type Channel = { id: ChannelId, server_id: ServerId | null, kind: ChannelKind, name: string | null, position: number, };